fix(both): fix file encryption type again

+++ android/src/main/java/com/margelo/nitro/nitroaes/NitroAes.kt
@@ -89,7 +89,7 @@ class NitroAes : HybridNitroAesSpec() {
     return Promise.async {
       ensureInitialized()
       val (auth, padding) = doEncryptFile(key, iv, hmacKey, inputPath, outputPath)
-      EncryptFileResult(auth, padding)
+      EncryptFileResult(auth, padding.toDouble())
     }
   }
